deps: Change Tween engine for Phaser's
All tweening engines I found are terrible documentation-wise and/or API-wise, I'm completely unsatisfied by this fact, so I took Phaser 3's tweening engine, stripped it down, and rewrote it to C++ with only the features we will use or potentially use.

And at least, this is documented and is easy to use.